move printing stuff to uci files
