+++ src/search.c
@@ -36,6 +36,7 @@
 #include "threads.h"
 #include "transposition.h"
 #include "syzygy.h"
+#include "uci.h"
 
 
 int Reductions[32][32];
@@ -63,72 +64,6 @@ static bool IsRepetition(const Position *pos) {
             return true;
 
     return false;
-}
-
-// Get ready to start a search
-static void PrepareSearch(Position *pos, Thread *threads) {
-
-    // Setup threads for a new search
-    for (int i = 0; i < threads->count; ++i) {
-        memset(&threads[i], 0, offsetof(Thread, pos));
-        memcpy(&threads[i].pos, pos, sizeof(Position));
-    }
-
-    // Mark TT as used
-    TT.dirty = true;
-}
-
-// Translates an internal mate score into distance to mate
-INLINE int MateScore(const int score) {
-    return score > 0 ?  ((MATE - score) / 2) + 1
-                     : -((MATE + score) / 2);
-}
-
-// Print thinking
-static void PrintThinking(const Thread *thread, int score, int alpha, int beta) {
-
-    // Determine whether we have a centipawn or mate score
-    char *type = abs(score) >= MATE_IN_MAX ? "mate" : "cp";
-
-    // Determine if score is an upper or lower bound
-    char *bound = score >= beta  ? " lowerbound"
-                : score <= alpha ? " upperbound"
-                                 : "";
-
-    // Translate internal score into printed score
-    score = abs(score) >=  MATE_IN_MAX ? MateScore(score)
-          : abs(score) >= TBWIN_IN_MAX ? score
-                                       : score * 100 / P_MG;
-
-    TimePoint elapsed = TimeSince(Limits.start);
-    Depth seldepth    = thread->seldepth;
-    uint64_t nodes    = TotalNodes(thread);
-    uint64_t tbhits   = TotalTBHits(thread);
-    int hashFull      = HashFull();
-    int nps           = (int)(1000 * nodes / (elapsed + 1));
-
-    // Basic info
-    printf("info depth %d seldepth %d score %s %d%s time %" PRId64
-           " nodes %" PRIu64 " nps %d tbhits %" PRIu64 " hashfull %d pv",
-            thread->depth, seldepth, type, score, bound, elapsed,
-            nodes, nps, tbhits, hashFull);
-
-    // Principal variation
-    for (int i = 0; i < thread->pv.length; i++)
-        printf(" %s", MoveToStr(thread->pv.line[i]));
-
-    printf("\n");
-    fflush(stdout);
-}
-
-// Print conclusion of search - best move and ponder move
-static void PrintConclusion(const Thread *thread) {
-
-    printf("bestmove %s", MoveToStr(thread->bestMove));
-    if (thread->ponderMove)
-        printf(" ponder %s", MoveToStr(thread->ponderMove));
-    printf("\n\n");
-    fflush(stdout);
 }
 
 INLINE bool PawnOn7th(const Position *pos) {
@@ -561,7 +496,7 @@ static int AspirationWindow(Thread *thread) {
 }
 
 // Iterative deepening
-void *IterativeDeepening(void *voidThread) {
+static void *IterativeDeepening(void *voidThread) {
 
     Thread *thread = voidThread;
     bool mainThread = thread->index == 0;
@@ -590,6 +525,19 @@ void *IterativeDeepening(void *voidThread) {
     }
 
     return NULL;
+}
+
+// Get ready to start a search
+static void PrepareSearch(Position *pos, Thread *threads) {
+
+    // Setup threads for a new search
+    for (int i = 0; i < threads->count; ++i) {
+        memset(&threads[i], 0, offsetof(Thread, pos));
+        memcpy(&threads[i].pos, pos, sizeof(Position));
+    }
+
+    // Mark TT as used
+    TT.dirty = true;
 }
 
 // Root of search
